server/requirements.txt
django
djangorestframework
django-filter
markdown
PyYAML
uritemplate
django-guardian
drf-yasg
django-cors-headers
graphene_django
grpcio
grpcio-tools
pytest-django